mask_detection_webapp_face_cascade/requirements.txt
gunicorn
Flask==3.0.2
Flask_Cors==4.0.0
Flask_SocketIO==5.3.6
numpy==1.26.4
Pillow==10.2.0
tensorflow==2.14.0

